add video player page

+++ app/src/main/java/com/king/app/coolg_kt/view/widget/video/OnVideoDurationListener.java
@@ -7,5 +7,5 @@ package com.king.app.coolg_kt.view.widget.video;
  * @date: 2020/9/29 9:11
  */
 public interface OnVideoDurationListener {
-    void onReceiveDuration(int duration);
+    void onReceiveDuration(long duration);
 }
